Add `useScroll` hook to abstract common logic in `KeyboardAwareFlatList`
